feat: temperature annealing + exp1.1 balancing loss sweep
Add router temperature annealing: linearly interpolate from
router_temp_start to router_temp_end over training steps. Updated via
eqx.tree_at each step (stored as jnp.float32 for dynamic updates).

Exp1.1 sweeps 5 balancing loss weights (0.005, 0.05, 0.1, 0.5, 1.0)
all with annealing 2.0→0.5, 3 seeds each. Motivated by exp1 showing
complete routing collapse (entropy=0, 100% layer 0).

+++ choosy/modeling/choosy.py
@@ -72,7 +72,8 @@ class ChoosyTransformer(eqx.Module):
         self.num_heads = num_heads
         self.max_seq_len = max_seq_len
         self.routing_loss_weight = routing_loss_weight
-        self.router_temperature = router_temperature
+        # Store as jnp scalar so it can be updated via eqx.tree_at for annealing
+        self.router_temperature = jnp.float32(router_temperature)
 
         keys = jr.split(key, num_pool_layers + 4)
